feat(ros2_entity_controller): finish sample
